set WARN lvl colours, suppress resulting logs
clean console output, removed the package path logging, added colours to [warn] [error] and [info] and added brackets.
added version.properties and modified script so that the Version only needs to be set in 1 place (pom.xml) for the script (not including readme). instead of in wowchat.conf and the pom.xml